refactor(evm): split the dispatch loop into std and zkevm partials
