+++ src/Nethermind/Nethermind.Config/BlocksConfig.cs
@@ -66,16 +66,11 @@ namespace Nethermind.Config
         public int MempoolPreWarmConcurrency { get; set; } = 0;
 
         public int BlockProductionTimeoutMs { get; set; } = 4_000;
-#if ZK_EVM
-        // No 0.25 initializer: storing a double constant is the only FPU instruction this ctor
-        // would contribute to the rv64ima guest image. This DELIBERATELY diverges from the
-        // DefaultValue = "0.25" documented on IBlocksConfig - the guest reads 0 instead. That is
-        // sound only because the sole consumer is block production, which never runs in the guest;
-        // anything else reading this in a guest build would silently get a different number rather
-        // than fail, so a new consumer must revisit this.
+
+        // The 0.25 default emits an FP constant load the guest's ISA gate rejects; only block production reads it.
         public double SingleBlockImprovementOfSlot { get; set; }
-#else
-        public double SingleBlockImprovementOfSlot { get; set; } = 0.25;
+#if !ZK_EVM
+            = 0.25;
 #endif
 
         public int GenesisTimeoutMs { get; set; } = 40_000;
