+++ src/Nethermind/Nethermind.Evm/ICodeCache.cs
@@ -8,8 +8,9 @@ namespace Nethermind.Evm;
 
 /// <summary>Bytecode cache keyed by code hash, used by <see cref="CacheCodeInfoRepository"/> to avoid re-reading and re-parsing code from the world state.</summary>
 /// <remarks>
-/// The witness/stateless path injects <see cref="NoopCodeCache"/> so that every code lookup goes through the world state and
-/// is captured in the generated witness; a process-wide cache hit would otherwise serve code without a world-state read. See <see cref="CodeInfoRepository"/>.
+/// Witness <em>generation</em> injects <see cref="NoopCodeCache"/> so that every code lookup goes through the world state and is
+/// captured in the witness. Stateless <em>validation</em> may cache, being keyed by code hash exactly as the witness stores code.
+/// See <see cref="CodeInfoRepository"/>.
 /// </remarks>
 public interface ICodeCache
 {
